add roles, role middleware, policy for user model

+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\User\GetUserRequest;
 use App\Http\Requests\User\UpdateUserRequest;
 use App\Http\Requests\User\RemoveUserRequest;
+use Illuminate\Support\Facades\Auth;
 use App\Models\User;
 use Illuminate\Support\Facades\Hash;
 
@@ -14,17 +15,28 @@ class UserController extends Controller
     {
         $validated = $request->validated();
 
-        return response()->json(User::findOrFail($validated['id']));
+        $user = User::findOrFail($validated['id']);
+
+        if (Auth::user()->cannot('get', $user)) {
+            return response(['message' => __('validation.custom.not_enough_rights')], 403);
+        }
+
+        return response()->json($user);
     }
 
     public function update(UpdateUserRequest $request)
     {
         $validated = $request->validated();
 
+        $user = User::findOrFail($validated['id']);
+
+        if (Auth::user()->cannot('update', $user)) {
+            return response(['message' => __('validation.custom.not_enough_rights')], 403);
+        }
+
         if (isset($validated['password']))
             $validated['password'] = Hash::make($validated['password']);
 
-        $user = User::findOrFail($validated['id']);
 
         $user->update($validated);
 
@@ -36,6 +48,10 @@ class UserController extends Controller
         $validated = $request->validated();
 
         $user = User::findOrFail($validated['id']);
+
+        if (Auth::user()->cannot('remove', $user)) {
+            return response(['message' => __('validation.custom.not_enough_rights')], 403);
+        }
 
         $user->delete();
